refactor: remove vercel cron dependency for document job processing
