+++ .env.example
@@ -27,10 +27,8 @@ PLATFORM_ROOT_HOSTS="localhost:3000"
 # Vercel Blob (required only if uploads are enabled)
 BLOB_READ_WRITE_TOKEN=""
 
-# Vercel cron security + queue processing controls
-# Set CRON_SECRET in Vercel for authorized cron invocations.
-CRON_SECRET=""
-PDF_CRON_BATCH_LIMIT="5"
+# Inline processing controls for document queue dispatch (Next.js route handlers)
+PDF_INLINE_BATCH_LIMIT="5"
 
 # Optional Postgres/Prisma tuning knobs
 # PG_POOL_MAX="10"
