+++ app/api/documents/render-jobs/process/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { errorResponse, successResponse, validateSession } from "@/lib/api-utils";
-import { processNextDocumentRenderJob } from "@/lib/documents/service";
+import { processDocumentRenderJobsBatch } from "@/lib/documents/service";
 
 function canRunWorker(role: string) {
   return role === "SUPERADMIN" || role === "MANAGER";
@@ -9,16 +9,8 @@ function canRunWorker(role: string) {
 export const runtime = "nodejs";
 export const maxDuration = 300;
 
-function isAuthorizedCronRequest(request: NextRequest): boolean {
-  const expectedSecret = process.env.CRON_SECRET?.trim();
-  if (!expectedSecret) return false;
-  const authorization = request.headers.get("authorization");
-  if (!authorization) return false;
-  return authorization === `Bearer ${expectedSecret}`;
-}
-
 function parseBatchLimit(request: NextRequest): number {
-  const defaultLimit = Number(process.env.PDF_CRON_BATCH_LIMIT ?? 5);
+  const defaultLimit = Number(process.env.PDF_INLINE_BATCH_LIMIT ?? 5);
   const { searchParams } = new URL(request.url);
   const requested = Number(searchParams.get("limit") ?? defaultLimit);
   if (!Number.isFinite(requested)) return 5;
@@ -26,37 +18,20 @@ function parseBatchLimit(request: NextRequest): number {
 }
 
 async function runProcessor(request: NextRequest) {
-  const cronAuthorized = isAuthorizedCronRequest(request);
+  const sessionResult = await validateSession(request);
+  if (sessionResult instanceof NextResponse) return sessionResult;
+  const { session } = sessionResult;
 
-  if (!cronAuthorized) {
-    const sessionResult = await validateSession(request);
-    if (sessionResult instanceof NextResponse) return sessionResult;
-    const { session } = sessionResult;
-
-    if (!canRunWorker(session.user.role)) {
-      return errorResponse("Only SUPERADMIN and MANAGER can process render jobs", 403);
-    }
+  if (!canRunWorker(session.user.role)) {
+    return errorResponse("Only SUPERADMIN and MANAGER can process render jobs", 403);
   }
 
   const limit = parseBatchLimit(request);
-  const results = [];
-  for (let index = 0; index < limit; index += 1) {
-    const result = await processNextDocumentRenderJob();
-    results.push(result);
-    if (!result.processed) break;
-  }
-
-  const processedCount = results.filter((row) => row.processed).length;
-  const failedCount = results.filter((row) => "status" in row && row.status === "FAILED").length;
-  const last = results.at(-1);
+  const batch = await processDocumentRenderJobsBatch(limit);
 
   return successResponse({
-    mode: cronAuthorized ? "cron" : "manual",
-    limit,
-    processedCount,
-    failedCount,
-    stopReason: last && !last.processed && "reason" in last ? last.reason : null,
-    results,
+    mode: "manual",
+    ...batch,
   });
 }
