--- vercel.json
@@ -1,9 +1,0 @@
-{
-  "$schema": "https://openapi.vercel.sh/vercel.json",
-  "crons": [
-    {
-      "path": "/api/documents/render-jobs/process",
-      "schedule": "*/2 * * * *"
-    }
-  ]
-}
